# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Zagraj ze Mną",
    "description": "API platrofmy Zagraj ze Mną"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "",
      "schemes": [
        "http"
      ]
    }
  ],
...
  "paths": {
...
    "/api/profile/getUserDetails": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Pobranie szczegółów użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi na pobranie szczegółów swojego profilu.",
        "operationId": "getUserDetails",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Szczegóły profilu użytkownika zostały pomyślnie pobrane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "jan_kowalski"
                    },
                    "username": {
                      "type": "string",
                      "example": "Johnny"
                    },
                    "about": {
                      "type": "string",
                      "example": "Pasjonat gier i programowania."
                    },
                    "country": {
                      "type": "string",
                      "example": "Polska"
                    },
                    "city": {
                      "type": "string",
                      "example": "Warszawa"
                    },
                    "contact": {
                      "type": "string",
                      "example": "jan@example.com"
                    },
                    "IDlanguage": {
                      "type": "integer",
                      "example": 1,
                      "description": "ID języka zgodnie z tabelą Languages"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - problem z danymi użytkownika."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/changePassword": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Zmiana hasła użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi na zmianę jego hasła.",
        "operationId": "changePassword",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Dane wymagane do zmiany hasła",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "Aktualne hasło użytkownika"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "Nowe hasło użytkownika"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "Potwierdzenie nowego hasła"
                  }
                },
                "required": [
                  "oldPassword",
                  "newPassword",
                  "confirmPassword"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Hasło zostało pomyślnie zmienione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password successfully changed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe dane wejściowe, hasła nie pasują do siebie lub nowe hasło jest takie samo jak stare"
          },
          "404": {
            "description": "Użytkownik nie został znaleziony"
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/postUsername": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Aktualizacja nazwy użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi aktualizować jego nazwę użytkownika.",
        "operationId": "updateUsername",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Żądanie aktualizacji nazwy użytkownika",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Nowa nazwa użytkownika"
                  }
                },
                "required": [
                  "username"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nazwa użytkownika została pomyślnie zaktualizowana",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Nazwa użytkownika została pomyślnie zaktualizowana."
                    },
                    "username": {
                      "type": "string",
                      "example": "nowy_nickname"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - brak nazwy użytkownika lub nazwa jest pusta."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "409": {
            "description": "Nazwa użytkownika jest już zajęta."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/updateAbout": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Aktualizacja informacji o użytkowniku",
        "description": "Pozwala zalogowanemu użytkownikowi aktualizować informacje o sobie w profilu.",
        "operationId": "updateAbout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Żądanie aktualizacji informacji o użytkowniku",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "about": {
                    "type": "string",
                    "description": "Informacje o użytkowniku do aktualizacji",
                    "example": "Pasjonat gier komputerowych i programowania."
                  }
                },
                "required": [
                  "about"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Informacje o użytkowniku zostały pomyślnie zaktualizowane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "About information updated successfully."
                    },
                    "about": {
                      "type": "string",
                      "example": "Pasjonat gier komputerowych i programowania."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - brak wymaganego pola 'about' lub pole jest puste."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/updateCountry": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Aktualizacja kraju użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi aktualizować kraj w jego profilu.",
        "operationId": "updateCountry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Żądanie aktualizacji kraju użytkownika",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "country": {
                    "type": "string",
                    "description": "Nazwa kraju do zapisania w profilu użytkownika",
                    "example": "Polska"
                  }
                },
                "required": [
                  "country"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kraj użytkownika został pomyślnie zaktualizowany",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Country updated successfully."
                    },
                    "country": {
                      "type": "string",
                      "example": "Polska"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - brak wymaganego pola 'country' lub pole jest puste."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/updateCity": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Aktualizacja miasta użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi aktualizować miasto w jego profilu.",
        "operationId": "updateCity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Żądanie aktualizacji miasta użytkownika",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "description": "Nazwa miasta do zapisania w profilu użytkownika",
                    "example": "Warszawa"
                  }
                },
                "required": [
                  "city"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Miasto użytkownika zostało pomyślnie zaktualizowane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "City updated successfully."
                    },
                    "city": {
                      "type": "string",
                      "example": "Warszawa"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - brak wymaganego pola 'city' lub pole jest puste."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/updateContact": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Aktualizacja informacji kontaktowej użytkownika",
        "description": "Pozwala zalogowanemu użytkownikowi aktualizować jego informacje kontaktowe w profilu.",
        "operationId": "updateContact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Żądanie aktualizacji informacji kontaktowej użytkownika",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contact": {
                    "type": "string",
                    "description": "Nowe informacje kontaktowe użytkownika",
                    "example": "jankowalski#1234"
                  }
                },
                "required": [
                  "contact"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Informacje kontaktowe użytkownika zostały pomyślnie zaktualizowane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact information updated successfully."
                    },
                    "contact": {
                      "type": "string",
                      "example": "jan.kowalski@example.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - brak wymaganego pola 'contact' lub pole jest puste."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/getAllLanguages": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Pobiera wszystkie języki",
        "description": "Zwraca listę wszystkich języków wraz z ich ID i nazwą.",
        "operationId": "getAllLanguages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista języków",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "language": {
                        "type": "string",
                        "example": "English"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nie znaleziono języków"
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
    "/api/profile/setUserLanguage": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Ustawienie języka użytkownika",
        "description": "Zezwala zalogowanemu użytkownikowi na wybór języka.",
        "operationId": "setUserLanguage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "ID języka do ustawienia dla użytkownika",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "languageId": {
                    "type": "integer",
                    "description": "ID języka, który ma zostać przypisany do użytkownika",
                    "example": 2
                  }
                },
                "required": [
                  "languageId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Język użytkownika został zaktualizowany"
          },
          "404": {
            "description": "Nie znaleziono języka lub użytkownika"
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    },
...
    "/api/profile/addGameToShelf": {

      "post": {
        "tags": [
          "Profil"
        ],

        "summary": "Dodaj grę do półki użytkownika",
        "description": "Umożliwia zalogowanemu użytkownikowi dodanie nowej gry do swojej półki.",
        "operationId": "addGameToShelf",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {

          "description": "Dane potrzebne do dodania gry do półki",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID_GAME": {
                    "type": "integer",
                    "example": 456,
                    "description": "ID gry, która ma być dodana do półki"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Gra została pomyślnie dodana do półki",

            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {

                    "message": {
                      "type": "string",
                      "example": "Gra została dodana do półki."
                    },
                    "shelfDetails": {
                      "type": "object",
                      "properties": {
                        "ID_SHELF": {
                          "type": "integer",
                          "example": 123,
                          "description": "ID nowo utworzonej półki"
                        },
                        "ID_USER": {
                          "type": "integer",
                          "example": 789,
                          "description": "ID użytkownika, który dodał grę"
                        },
                        "ID_GAME": {
                          "type": "integer",
                          "example": 456,
                          "description": "ID gry dodanej do półki"
                        }
                      }

                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - problem z przesłanymi danymi."
          },
          "403": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "404": {
            "description": "Nie znaleziono żądanej gry."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera."
          }
        }
      }

    },
    "/api/profile/removeGameFromShelf": {
      "delete": {
        "tags": [
          "Profil"
        ],
        "summary": "Usuń grę z półki użytkownika",
        "description": "Usunięcie gry z półki zalogowanego użytkownika na podstawie podanego ID_GAME.",
        "operationId": "removeGameFromShelf",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID_GAME": {
                    "type": "integer",
                    "example": 456,
                    "description": "ID gry, która ma być usunięta z półki"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gra została pomyślnie usunięta z półki",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Gra została usunięta z półki."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - problem z danymi wejściowymi"
          },
          "403": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy"
          },
          "404": {
            "description": "Nie znaleziono gry na półce tego użytkownika"
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }

    },
    "/api/profile/getUserStats": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Pobranie statystyk użytkownika",
        "description": "Zwraca ilość gier na półce użytkownika oraz ilość lobby, do których jest przypisany.",
        "operationId": "getUserStats",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statystyki użytkownika zostały pomyślnie pobrane",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "gamesOnShelf": {
                      "type": "integer",
                      "example": 5,
                      "description": "Ilość gier na półce użytkownika"
                    },
                    "lobbiesJoined": {
                      "type": "integer",
                      "example": 3,
                      "description": "Ilość lobby, do których użytkownik jest przypisany"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe żądanie - problem z żądaniem od klienta."
          },
          "401": {
            "description": "Nieautoryzowane - brak tokena lub token jest nieprawidłowy."
          },
          "403": {
            "description": "Użytkownik nie znaleziony."
          },
          "500": {
            "description": "Wewnętrzny błąd serwera"
          }
        }
      }
    }    
  },  
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}